add gtpv1.IP() that returns net.IP
